Extract 'fetch_sources' to separate rakefile
'combine_sources' is about to be renamed to make it more clear that it's
only about making the `merged.csv` file, and not working with any of the
other non Membership-related sources. To help with that, first extract
'fetch_sources' as a separate namespace.
